+++ src/main/java/com/mastery/java/task/service/EmployeeService.java
@@ -8,6 +8,8 @@ public interface EmployeeService {
     public List<Employee> getAllEmployees();
 
     public void saveEmployee(Employee employee);
-    public Employee getEmployee(int id);
-    public void deleteEmployee(int id);
+
+    public Employee getEmployee(long id);
+
+    public void deleteEmployee(long id);
 }
