+++ src/main/java/com/mastery/java/task/service/EmployeeServiceImpl.java
@@ -5,6 +5,7 @@ import com.mastery.java.task.dao.EmployeeDao;
 import com.mastery.java.task.dto.Employee;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -25,18 +26,17 @@ public class EmployeeServiceImpl implements EmployeeService {
     }
 
     @Override
-    public Employee getEmployee(int id) {
+    public Employee getEmployee(long id) {
         Employee employee = null;
         Optional<Employee> optional = employeeDao.findById(id);
-        if(optional.isPresent())
-        {
+        if (optional.isPresent()) {
             employee = optional.get();
         }
         return employee;
     }
 
     @Override
-    public void deleteEmployee(int id){
+    public void deleteEmployee(long id) {
         employeeDao.deleteById(id);
     }
